requirements.txt
python-dotenv
openai==0.28
PyMUPDF
python-pptx